homedrive/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
argon2-cffi==23.1.0
certbot==2.7.4
